Added ID accept/deny system
When a patron reaches the counter they hand over their ID. A variable of it being real or fake is added to the front of an array. When accepting or denying the end of the array is checked (earliest patron) to determine what happens (incomplete. Need text + potion system integration) and the value at the end of the array is removed.